Move set_default_with_acquisition() to opengever.base.acquisition.
