+++ opengever/base/behaviors/utils.py
@@ -123,34 +123,6 @@ class RestrictedVocabularyFactory(object):
         return self.field.default
 
 
-def set_default_with_acquisition(field, default=None):
-    """
-    Sets a default value generator which uses the value
-    from the parent object, if existing, otherwise it uses
-    the given default value.
-    """
-
-    def default_value_generator(data):
-        container = data.context
-
-        acquired_value = acquire_field_value(field, container)
-        if acquired_value is not NO_VALUE_FOUND:
-            return acquired_value
-
-        # otherwise use default value
-        if default:
-            # XXX: Use sentinel value (Issue #2029)
-            return default
-        else:
-            # use first value
-            try:
-                return tuple(data.widget.terms)[0].value
-            except AttributeError:
-                return None
-
-    return default_value_generator
-
-
 def propagate_vocab_restrictions(container, event, restricted_fields, marker):
 
     def dottedname(field):
